AI Query API with Prompt Control/ai-query-api/requirements.txt
fastapi
uvicorn
python-dotenv
requests
PyMuPDF
sentence-transformers
chromadb
jinja2
